% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_barcode.R
\name{consulta_barcode}
\alias{consulta_barcode}
\title{Title}
\usage{
consulta_barcode(apikey, barCode, .dias = 1, .latitude = -9.6432331,
  .longitude = -35.7190686, .raio = 15)
}
\arguments{
\item{apikey}{- Chave do API, obtida mediante pedido a api [at] sefaz.al.gov.br}

\item{barCode}{- String com o codigo de barras}

\item{.dias}{- Numero de dias. Caso não fornecido, e apenas um dia}

\item{.latitude}{- Latitude do centro do circulo onde a query vai investigar (formato Lat-Lon, nao eh UTM nem outros formatos). Existe um default, mas nao eh relevante.}

\item{.longitude}{- Lognitude do centro do circulo onde a query vai investigar (formato Lat-Lon, nao eh UTM nem outros formatos). Existe um default, mas nao eh relevante.}

\item{.raio}{- Raio do circulo. Default eh 15 kms}
}
\value{
- Retorna um dataframe com os seguintes objetos:
codGetin	String -- Codigo do produto.
dscProduto	String	-- Descricao do produto.
dthEmissaoUltimaVenda	String -- Data de Ultima venda do produto.
valUnitarioUltimaVenda	Double -- Valor unitário da Ultima venda.
valUltimaVenda	Double	-- Valor da Ultima venda.
valMinimoVendido	Double	-- Valor minimo vendido no periodo pesquisado.
valMaximoVendido	Double	-- Valor maximo vendido no periodo pesquisado.
txtDataUltimaEmissao	String	-- Calculo do tempo da Data da ultima emissao do produto.
numCNPJ	String	-- Numero Cadastro Nacional da pessoa juridica.
nomRazaoSocial	String	-- Nome da razao social do contribuinte.
nomFantasia	String	-- Nome fantasia do contribuinte.
numTelefone	String	-- Numero de telefone do contribuinte.
nomLogradouro	String	-- Nome do logradouro do estabelecimento.
numImovel	String	-- Numero do estabelecimento.
nomBairro	String	-- Nome do bairro do estabelecimento.
numCep	String	-- CEP.
nomMunicipio	String -- Municipio.
numLatitude	Double	-- Latitude do estabelecimento.
numLongitude	Double	-- Longitude do estabelecimento.
codNcm	Long	-- Codigo NCM do produto.
}
\description{
Title
}
\examples{
consulta_barcode(apikey=key,barCode="07894900010015")

}
