% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_min_preco.R
\name{consulta_min_preco}
\alias{consulta_min_preco}
\title{Consulta o menor preco em um determinado estabelecimento por codigo de barras}
\usage{
consulta_min_preco(apikey, .cnpj, .codigoBarras, .quantidadeDeDias = 3)
}
\arguments{
\item{apikey}{- Chave do API, obtida mediante pedido a api [at] sefaz.al.gov.br}

\item{.cnpj}{- String    -- Numero Cadastro Nacional da pessoa juridica.}

\item{.codigoBarras}{- String com o codigo de barras}

\item{.quantidadeDeDias}{- Numero de dias. Default (caso missing)=3}
}
\value{
- Retorna um dataframe com os seguintes objetos:
codGetin	String -- Codigo do produto.
dscProduto	String	-- Descricao do produto.
dthEmissaoUltimaVenda	String -- Data de Ultima venda do produto.
valUnitarioUltimaVenda	Double -- Valor unitário da Ultima venda.
valUltimaVenda	Double	-- Valor da Ultima venda.
valMinimoVendido	Double	-- Valor minimo vendido no periodo pesquisado.
valMaximoVendido	Double	-- Valor maximo vendido no periodo pesquisado.
txtDataUltimaEmissao	String	-- Calculo do tempo da Data da ultima emissao do produto.
numCNPJ	String	-- Numero Cadastro Nacional da pessoa juridica.
nomRazaoSocial	String	-- Nome da razao social do contribuinte.
nomFantasia	String	-- Nome fantasia do contribuinte.
numTelefone	String	-- Numero de telefone do contribuinte.
nomLogradouro	String	-- Nome do logradouro do estabelecimento.
numImovel	String	-- Numero do estabelecimento.
nomBairro	String	-- Nome do bairro do estabelecimento.
numCep	String	-- CEP.
nomMunicipio	String -- Municipio.
numLatitude	Double	-- Latitude do estabelecimento.
numLongitude	Double	-- Longitude do estabelecimento.
codNcm	Long	-- Codigo NCM do produto.
}
\description{
Consulta o menor preco em um determinado estabelecimento por codigo de barras
}
\examples{
consulta_min_preco(apikey=key,.cnpj="08618647000355",.codigoBarras =  "7898286201968",.quantidadeDeDias = 1)
}
